migration: db:gen add telegram star payment type
