+++ mini-app/drizzle/0010_lucky_stature.sql
@@ -1,0 +1,1 @@
+ALTER TYPE "payment_types" ADD VALUE 'STAR';
