+++ mini-app/drizzle/meta/_journal.json
@@ -71,6 +71,13 @@
       "when": 1740295510278,
       "tag": "0009_eager_chronomancer",
       "breakpoints": true
+    },
+    {
+      "idx": 10,
+      "version": "7",
+      "when": 1740602406532,
+      "tag": "0010_lucky_stature",
+      "breakpoints": true
     }
   ]
 }
