+++ mini-app/src/db/enum.ts
@@ -59,3 +59,4 @@ export type NotificationType = (typeof notificationType.enumValues)[number];
 export type NotificationStatus = (typeof notificationStatus.enumValues)[number];
 export type NotificationItemType = (typeof notificationItemType.enumValues)[number];
 export type EventPoaResultStatus = (typeof eventPoaResultStatus.enumValues)[number];
+export type PaymentTypes = (typeof paymentTypes.enumValues)[number];
