+++ mini-app/src/server/db/orders.db.ts
@@ -95,6 +95,18 @@ async function checkIfSoldOut(event_uuid: string, ticketOrderType: OrderTypeValu
   return { isSoldOut: TicketsCount[0].ticket_count >= capacity, soldCount: TicketsCount[0].ticket_count };
 }
 
+/**
+ * Check if there's already a completed order for this event, user, and orderType.
+ */
+const findExistingCompletedOrder = async (eventUuid: string, telegramUserId: number, orderType: OrderTypeValues) =>
+  db.query.orders.findFirst({
+    where: and(
+      eq(orders.user_id, telegramUserId),
+      eq(orders.order_type, orderType),
+      eq(orders.event_uuid, eventUuid),
+      eq(orders.state, "completed")
+    ),
+  });
 const ordersDB = {
   getEventOrders,
   updateOrderState,
@@ -102,6 +114,7 @@ const ordersDB = {
   createPromoteToOrganizerOrder,
   getPromoteToOrganizerOrder,
   checkIfSoldOut,
+  findExistingCompletedOrder,
 };
 
 export default ordersDB;
